+++ pkgs/dart_tooling_mcp_server/lib/src/server.dart
@@ -20,6 +20,7 @@ final class DartToolingMCPServer extends MCPServer
     with
         LoggingSupport,
         ToolsSupport,
+        RootsTrackingSupport,
         DartAnalyzerSupport,
         DartCliSupport,
         PubSupport,